Extract task execution to TaskRunner
At its core a process is a function that accepts an input, executes a
task and returns an asynchronous result. TaskProcessor implements this
functionality together with the dataflow interface of the process, so
that a task can only be executed by feeding values through the process
input channels.

Move the execution of a task to a separate `TaskRunner` class:

- the resolution of the task inputs and outputs
- the store directory and cache lookup
- the delegation to the executor, including task arrays
- the task error strategy and retries

TaskProcessor keeps the dataflow interface -- the operator, the process
state, the output binding and the publishing -- as well as the process
scoped services shared by all task executions, such as the name, the
config, the executor and the task environment.

The processor creates a runner with `taskCompleted` as the completion
handler, so that the runner does not depend on the dataflow interface:
a task can be executed by invoking `TaskRunner.submit()` directly,
without the process being part of a dataflow network (see the test
'should execute a task without a dataflow network').

Note that `resumeOrDie` is now synchronized on the runner rather than on
the processor, so it no longer excludes `bindOutputs` and
`sendPoisonPill`. These methods do not share mutable state, the coarse
lock was incidental.
