+++ modules/nextflow/src/main/groovy/nextflow/processor/TaskPollingMonitor.groovy
@@ -677,7 +677,7 @@ class TaskPollingMonitor implements TaskMonitor {
             if (evict(handler)) {
                 handler.decProcessForks()
             }
-            fault = handler.task.processor.resumeOrDie(handler?.task, error, handler.getTraceRecord())
+            fault = handler.task.processor.getRunner().resumeOrDie(handler?.task, error, handler.getTraceRecord())
             log.trace "Task fault (1): $fault"
         }
         finally {
@@ -743,7 +743,7 @@ class TaskPollingMonitor implements TaskMonitor {
 
     protected void finalizeTask( TaskHandler handler ) {
         // finalize the task execution
-        final fault = handler.task.processor.finalizeTask(handler)
+        final fault = handler.task.processor.getRunner().finalizeTask(handler)
 
         // notify task completion
         session.notifyTaskComplete(handler)
