+++ modules/nextflow/src/testFixtures/groovy/test/ScriptHelper.groovy
@@ -347,7 +347,7 @@ class MockTaskHandler extends TaskHandler {
             task.code.call()
         }
         status = TaskStatus.COMPLETED
-        final fault = task.processor.finalizeTask(this)
+        final fault = task.processor.getRunner().finalizeTask(this)
         if( fault instanceof TaskFault )
             task.processor.session.fault(fault)
     }
